stmhal: Rename mphal.* to stmhal_mphal.* to be consistent with other ports.
The idea is that "mphal.h" will be central header which describes standard
MPHAL API, consistent across ports.
