+++ stmhal/stmhal_mphal.c
@@ -4,7 +4,7 @@
 #include "py/mpstate.h"
 #include "usb.h"
 #include "uart.h"
-#include "mphal.h"
+#include "stmhal_mphal.h"
 
 // this table converts from HAL_StatusTypeDef to POSIX errno
 const byte mp_hal_status_to_errno_table[4] = {
