+++ stmhal/stmhal_mphal.h
